Consolidate a lot of the functionality from the migrate_registration_responses management command into osf/utils/registrations to be used when flattening registration_metadata as part of supporting both workflows.
